requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-genai>=1.0.0
psycopg2-binary>=2.9.0
stripe>=8.0.0
resend>=2.0.0
